requirements.txt
python-dotenv
google-auth
google-auth-oauthlib
google-api-python-client
python-telegram-bot
pandas
openai
httpx
apscheduler
psycopg2-binary
